requirements.txt
Flask
python-dotenv
requests
pymongo
BeautifulSoup4
flask_cors